feature : 연극 상세 조회 기능 추가
Close #20

+++ src/main/java/com/chbb/theaketing/feature/drama/dto/DramaDateTimeDto.java
@@ -33,7 +33,7 @@ public class DramaDateTimeDto {
 
         @Schema(description = "남은 좌석")
         @NonNull
-        protected Integer remainSeats;
+        protected Long remainSeats;
 
     }
 }
