+++ src/main/java/com/chbb/theaketing/feature/payment/dto/PaymentDto.java
@@ -16,11 +16,11 @@ public class PaymentDto {
 
         @Schema(description = "예매 id", requiredMode = RequiredMode.NOT_REQUIRED)
         @Min(value = 1, message = "예매 정보가 올바르지 않습니다")
-        protected Integer reservationId;
+        protected Long reservationId;
 
         @Schema(description = "가격", requiredMode = RequiredMode.NOT_REQUIRED)
         @Min(value = 1, message = "결제 정보가 올바르지 않습니다")
-        protected Integer price;
+        protected Long price;
     }
 
 }
